Time: 5 ms (34.15%), Space: 6.3 MB (23.89%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -19,6 +19,6 @@ public:
             }
         }
                 
-        return static_cast<int>(floor(end));
+        return floor(end);
     }
 };
